refactor: share comment scanners across project detection and mapping
